+++ processor/src/main/java/arez/processor/GeneratorUtil.java
@@ -20,8 +20,8 @@ final class GeneratorUtil
   static final ClassName NULLABLE_CLASSNAME = ClassName.get( "javax.annotation", "Nullable" );
   static final ClassName INJECT_CLASSNAME = ClassName.get( "javax.inject", "Inject" );
   static final ClassName SINGLETON_CLASSNAME = ClassName.get( "javax.inject", "Singleton" );
+  static final ClassName DAGGER_BINDS_CLASSNAME = ClassName.get( "dagger", "Binds" );
   static final ClassName DAGGER_MODULE_CLASSNAME = ClassName.get( "dagger", "Module" );
-  static final ClassName DAGGER_PROVIDES_CLASSNAME = ClassName.get( "dagger", "Provides" );
   static final ClassName GUARDS_CLASSNAME = ClassName.get( "org.realityforge.braincheck", "Guards" );
   static final ClassName AREZ_CLASSNAME = ClassName.get( "arez", "Arez" );
   static final ClassName FLAGS_CLASSNAME = ClassName.get( "arez", "Flags" );