+++ processor/src/test/resources/expected/ImplicitSingletonModelDaggerModule.java
@@ -1,16 +1,12 @@
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface ImplicitSingletonModelDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static ImplicitSingletonModel provideComponent(final Arez_ImplicitSingletonModel component) {
-    return component;
-  }
+  ImplicitSingletonModel bindComponent(Arez_ImplicitSingletonModel component);
 }