+++ processor/src/test/resources/expected/SingletonWithIdModelDaggerModule.java
@@ -1,16 +1,12 @@
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface SingletonWithIdModelDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static SingletonWithIdModel provideComponent(final Arez_SingletonWithIdModel component) {
-    return component;
-  }
+  SingletonWithIdModel bindComponent(Arez_SingletonWithIdModel component);
 }