+++ processor/src/test/resources/expected/com/example/repository/CompleteRepositoryExampleRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface CompleteRepositoryExampleRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static CompleteRepositoryExampleRepository provideComponent(final Arez_CompleteRepositoryExampleRepository component) {
-    return component;
-  }
+  CompleteRepositoryExampleRepository bindComponent(Arez_CompleteRepositoryExampleRepository component);
 }