+++ processor/src/test/resources/expected/com/example/repository/DaggerEnabledRepositoryRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface DaggerEnabledRepositoryRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static DaggerEnabledRepositoryRepository provideComponent(final Arez_DaggerEnabledRepositoryRepository component) {
-    return component;
-  }
+  DaggerEnabledRepositoryRepository bindComponent(Arez_DaggerEnabledRepositoryRepository component);
 }