+++ processor/src/test/resources/expected/com/example/repository/InjectEnabledRepositoryRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface InjectEnabledRepositoryRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static InjectEnabledRepositoryRepository provideComponent(final Arez_InjectEnabledRepositoryRepository component) {
-    return component;
-  }
+  InjectEnabledRepositoryRepository bindComponent(Arez_InjectEnabledRepositoryRepository component);
 }