+++ processor/src/test/resources/expected/com/example/repository/NestedModel_BasicActionModelRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface NestedModel_BasicActionModelRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static NestedModel_BasicActionModelRepository provideComponent(final Arez_NestedModel_BasicActionModelRepository component) {
-    return component;
-  }
+  NestedModel_BasicActionModelRepository bindComponent(Arez_NestedModel_BasicActionModelRepository component);
 }