+++ processor/src/test/resources/expected/com/example/repository/PackageAccessRepositoryExampleRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface PackageAccessRepositoryExampleRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static PackageAccessRepositoryExampleRepository provideComponent(final Arez_PackageAccessRepositoryExampleRepository component) {
-    return component;
-  }
+  PackageAccessRepositoryExampleRepository bindComponent(Arez_PackageAccessRepositoryExampleRepository component);
 }