+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithCreateOnlyRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface RepositoryWithCreateOnlyRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static RepositoryWithCreateOnlyRepository provideComponent(final Arez_RepositoryWithCreateOnlyRepository component) {
-    return component;
-  }
+  RepositoryWithCreateOnlyRepository bindComponent(Arez_RepositoryWithCreateOnlyRepository component);
 }