+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithDetachNoneRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface RepositoryWithDetachNoneRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static RepositoryWithDetachNoneRepository provideComponent(final Arez_RepositoryWithDetachNoneRepository component) {
-    return component;
-  }
+  RepositoryWithDetachNoneRepository bindComponent(Arez_RepositoryWithDetachNoneRepository component);
 }