+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithExplicitNonStandardIdRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface RepositoryWithExplicitNonStandardIdRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static RepositoryWithExplicitNonStandardIdRepository provideComponent(final Arez_RepositoryWithExplicitNonStandardIdRepository component) {
-    return component;
-  }
+  RepositoryWithExplicitNonStandardIdRepository bindComponent(Arez_RepositoryWithExplicitNonStandardIdRepository component);
 }