+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithImplicitIdRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface RepositoryWithImplicitIdRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static RepositoryWithImplicitIdRepository provideComponent(final Arez_RepositoryWithImplicitIdRepository component) {
-    return component;
-  }
+  RepositoryWithImplicitIdRepository bindComponent(Arez_RepositoryWithImplicitIdRepository component);
 }