+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithInitializerNameCollisionModelRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface RepositoryWithInitializerNameCollisionModelRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static RepositoryWithInitializerNameCollisionModelRepository provideComponent(final Arez_RepositoryWithInitializerNameCollisionModelRepository component) {
-    return component;
-  }
+  RepositoryWithInitializerNameCollisionModelRepository bindComponent(Arez_RepositoryWithInitializerNameCollisionModelRepository component);
 }