+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithMultipleCtorsRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface RepositoryWithMultipleCtorsRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static RepositoryWithMultipleCtorsRepository provideComponent(final Arez_RepositoryWithMultipleCtorsRepository component) {
-    return component;
-  }
+  RepositoryWithMultipleCtorsRepository bindComponent(Arez_RepositoryWithMultipleCtorsRepository component);
 }