+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithProtectedConstructorRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface RepositoryWithProtectedConstructorRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static RepositoryWithProtectedConstructorRepository provideComponent(final Arez_RepositoryWithProtectedConstructorRepository component) {
-    return component;
-  }
+  RepositoryWithProtectedConstructorRepository bindComponent(Arez_RepositoryWithProtectedConstructorRepository component);
 }