+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithSingletonRepositoryDaggerModule.java
@@ -1,18 +1,14 @@
 package com.example.repository;
 
+import dagger.Binds;
 import dagger.Module;
-import dagger.Provides;
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
 import javax.inject.Singleton;
 
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface RepositoryWithSingletonRepositoryDaggerModule {
-  @Nonnull
-  @Provides
+  @Binds
   @Singleton
-  static RepositoryWithSingletonRepository provideComponent(final Arez_RepositoryWithSingletonRepository component) {
-    return component;
-  }
+  RepositoryWithSingletonRepository bindComponent(Arez_RepositoryWithSingletonRepository component);
 }